--- src/components/technologies/table.tsx
@@ -1,559 +1,0 @@
-import {
-  DndContext,
-  type DragEndEvent,
-  KeyboardSensor,
-  MouseSensor,
-  TouchSensor,
-  type UniqueIdentifier,
-  closestCenter,
-  useSensor,
-  useSensors,
-} from "@dnd-kit/core";
-import { restrictToVerticalAxis } from "@dnd-kit/modifiers";
-import { SortableContext, arrayMove, useSortable, verticalListSortingStrategy } from "@dnd-kit/sortable";
-import { CSS } from "@dnd-kit/utilities";
-import {
-  IconChevronDown,
-  IconChevronLeft,
-  IconChevronRight,
-  IconChevronUp,
-  IconChevronsLeft,
-  IconChevronsRight,
-  IconCircleCheckFilled,
-  IconDotsVertical,
-  IconGripVertical,
-  IconLayoutColumns,
-  IconLoader,
-  IconPlus,
-} from "@tabler/icons-react";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import {
-  ColumnDef,
-  Row,
-  SortingState,
-  VisibilityState,
-  flexRender,
-  getCoreRowModel,
-  getFacetedRowModel,
-  getFacetedUniqueValues,
-  getFilteredRowModel,
-  useReactTable,
-} from "@tanstack/react-table";
-import * as React from "react";
-import { useAuth } from "react-oidc-context";
-import { Link } from "react-router";
-import { toast } from "sonner";
-import { z } from "zod";
-
-import { Badge } from "@/ui/badge";
-import { Button } from "@/ui/button";
-import { Checkbox } from "@/ui/checkbox";
-import {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/ui/dropdown-menu";
-import { Label } from "@/ui/label";
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/ui/select";
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/ui/table";
-import { Tabs, TabsContent } from "@/ui/tabs";
-
-import { API_URL } from "@/constants/application";
-import { DELETE_TECHNOLOGY, GET_TECHNOLOGIES } from "@/constants/query-keys";
-
-import { technologySchema } from "@/schemas/technology";
-
-import { FilterInput } from "@/components/filter-input";
-
-import { useIsMobile } from "@/hooks/use-mobile";
-
-// Create a separate component for the drag handle
-function DragHandle({ id }: { id: number }) {
-  const { attributes, listeners } = useSortable({
-    id,
-  });
-
-  return (
-    <Button
-      {...attributes}
-      {...listeners}
-      variant="ghost"
-      size="icon"
-      className="text-muted-foreground size-7 hover:bg-transparent"
-    >
-      <IconGripVertical className="text-muted-foreground size-3" />
-      <span className="sr-only">Drag to reorder</span>
-    </Button>
-  );
-}
-
-function DraggableRow({ row }: { row: Row<z.infer<typeof technologySchema>> }) {
-  const { transform, transition, setNodeRef, isDragging } = useSortable({
-    id: row.original.id,
-  });
-
-  return (
-    <TableRow
-      data-state={row.getIsSelected() && "selected"}
-      data-dragging={isDragging}
-      ref={setNodeRef}
-      className="relative z-0 data-[dragging=true]:z-10 data-[dragging=true]:opacity-80"
-      style={{
-        transform: CSS.Transform.toString(transform),
-        transition: transition,
-      }}
-    >
-      {row.getVisibleCells().map((cell) => (
-        <TableCell key={cell.id}>{flexRender(cell.column.columnDef.cell, cell.getContext())}</TableCell>
-      ))}
-    </TableRow>
-  );
-}
-
-export const TechnologyTable = ({
-  data,
-  handlePagination,
-  handleSorting,
-  handleFilter,
-  isLoading = false,
-  pageSize,
-  rowCount,
-  pageIndex,
-}: {
-  data: z.infer<typeof technologySchema>[];
-  handlePagination?: (page: number, size: number) => void;
-  handleSorting?: (id: string, desc: "asc" | "desc") => void;
-  handleFilter?: (value: string) => void;
-  pageSize: number;
-  isLoading: boolean;
-  rowCount: number;
-  pageIndex: number;
-}) => {
-  const auth = useAuth();
-  const isMobile = useIsMobile();
-  const queryClient = useQueryClient();
-  const [localData, setLocalData] = React.useState(data);
-  const [rowSelection, setRowSelection] = React.useState({});
-  const [columnVisibility, setColumnVisibility] = React.useState<VisibilityState>({});
-
-  const [sorting, setSorting] = React.useState<SortingState>([]);
-  const [pagination, setPagination] = React.useState({
-    pageIndex: pageIndex,
-    pageSize: pageSize || 10,
-  });
-  const sortableId = React.useId();
-  const sensors = useSensors(useSensor(MouseSensor, {}), useSensor(TouchSensor, {}), useSensor(KeyboardSensor, {}));
-  const dataIds = React.useMemo<UniqueIdentifier[]>(() => data?.map(({ id }) => id) || [], [data]);
-
-  const columns: ColumnDef<z.infer<typeof technologySchema>>[] = [
-    {
-      id: "drag",
-      header: () => null,
-      cell: ({ row }) => <DragHandle id={row.original.id} />,
-    },
-    {
-      id: "select",
-      header: ({ table }) => (
-        <div className="flex items-center justify-center ">
-          <Checkbox
-            checked={table.getIsAllPageRowsSelected() || (table.getIsSomePageRowsSelected() && "indeterminate")}
-            onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
-            aria-label="Select all"
-          />
-        </div>
-      ),
-      cell: ({ row }) => (
-        <div className="flex items-center justify-center">
-          <Checkbox
-            checked={row.getIsSelected()}
-            onCheckedChange={(value) => row.toggleSelected(!!value)}
-            aria-label="Select row"
-          />
-        </div>
-      ),
-      enableSorting: false,
-      enableHiding: false,
-    },
-    {
-      accessorKey: "title",
-      header: "Title",
-      cell: ({ row }) => <div className="w-32">{row.original.title}</div>,
-      enableHiding: false,
-    },
-    {
-      accessorKey: "website",
-      header: "Website",
-      cell: ({ row }) => <div className="w-32">{row.original.website}</div>,
-    },
-    {
-      accessorKey: "moved",
-      header: "Moved",
-      cell: ({ row }) => (
-        <div className="w-32">
-          <Badge variant="outline" className="text-muted-foreground px-1.5">
-            {row.original.moved}
-          </Badge>
-        </div>
-      ),
-    },
-    {
-      accessorKey: "active",
-      header: "Active",
-      cell: ({ row }) => (
-        <Badge variant="outline" className="text-muted-foreground px-1.5">
-          {row.original.active ? (
-            <IconCircleCheckFilled className="fill-green-500 dark:fill-green-400" />
-          ) : (
-            <IconLoader />
-          )}
-          {row.original.active}
-        </Badge>
-      ),
-    },
-    {
-      id: "actions",
-      cell: ({ row }) => (
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button
-              variant="ghost"
-              className="data-[state=open]:bg-muted text-muted-foreground flex size-8"
-              size="icon"
-            >
-              <IconDotsVertical />
-              <span className="sr-only">Open menu</span>
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent align="end" className="w-32">
-            <Link to={`/technologies/edit/${row.id}`}>
-              <DropdownMenuItem className="cursor-pointer">Edit</DropdownMenuItem>
-            </Link>
-            <DropdownMenuSeparator />
-            <DropdownMenuItem variant="destructive" onClick={() => deleteTechnology(row.id)}>
-              Delete
-            </DropdownMenuItem>
-          </DropdownMenuContent>
-        </DropdownMenu>
-      ),
-    },
-  ];
-
-  const { mutate: deleteTechnology } = useMutation({
-    mutationFn: async (rowId: string) => {
-      await fetch(`${API_URL}/technologies/${rowId}`, {
-        method: "DELETE",
-        headers: {
-          "Content-type": "application/json",
-          Authorization: `Bearer ${auth.user?.access_token}`,
-        },
-      });
-    },
-    mutationKey: [DELETE_TECHNOLOGY],
-    onSuccess() {
-      queryClient.invalidateQueries({ queryKey: [GET_TECHNOLOGIES] });
-      toast.success("Technology has been deleted successfully");
-    },
-    onError(error) {
-      toast.error("Error deleting technology", {
-        description: error.message,
-      });
-    },
-  });
-
-  const defaultData = React.useMemo(() => [], []);
-
-  React.useEffect(() => {
-    setColumnVisibility((prev) => ({ ...prev, website: !isMobile, moved: !isMobile, active: !isMobile }));
-  }, [isMobile]);
-
-  React.useEffect(() => {
-    if (!isLoading && data) setLocalData(data);
-  }, [isLoading, data]);
-
-  const table = useReactTable({
-    data: localData ?? defaultData,
-    columns,
-    state: {
-      sorting,
-      columnVisibility,
-      rowSelection,
-      pagination,
-    },
-    rowCount: rowCount,
-    manualPagination: true,
-    manualSorting: true,
-    manualFiltering: true,
-    getRowId: (row) => row.id.toString(),
-    onRowSelectionChange: setRowSelection,
-    onSortingChange: (updaterOrValue) => {
-      const newSorting = typeof updaterOrValue === "function" ? updaterOrValue(sorting) : updaterOrValue;
-      if (handleSorting) {
-        const sort = newSorting[0];
-        if (newSorting.length > 0) handleSorting(sort.id, sort.desc ? "desc" : "asc");
-        else handleSorting("title", "desc");
-      }
-      return setSorting(newSorting);
-    },
-    onColumnVisibilityChange: setColumnVisibility,
-    onPaginationChange: (updaterOrValue) => {
-      const newPagination = typeof updaterOrValue === "function" ? updaterOrValue(pagination) : updaterOrValue;
-      if (handlePagination) {
-        handlePagination(newPagination.pageIndex, newPagination.pageSize);
-      }
-      return setPagination(newPagination);
-    },
-    getCoreRowModel: getCoreRowModel(),
-    getFilteredRowModel: getFilteredRowModel(),
-    getFacetedRowModel: getFacetedRowModel(),
-    getFacetedUniqueValues: getFacetedUniqueValues(),
-  });
-
-  function handleDragEnd(event: DragEndEvent) {
-    const { active, over } = event;
-    if (active && over && active.id !== over.id) {
-      setLocalData((data) => {
-        const oldIndex = dataIds.indexOf(active.id);
-        const newIndex = dataIds.indexOf(over.id);
-        return arrayMove(data, oldIndex, newIndex);
-      });
-    }
-  }
-
-  const handleDeleteSelectedTechnologies = () => {
-    const deleteArr = Object.keys(rowSelection);
-    for (const item of deleteArr) {
-      deleteTechnology(item);
-    }
-  };
-
-  return (
-    <Tabs defaultValue="outline" className="w-full flex-col justify-start gap-6">
-      <div className="flex items-center justify-between px-4 lg:px-6">
-        <div className="pr-2 max-w-72 w-72">{handleFilter ? <FilterInput handleFilter={handleFilter} /> : null}</div>
-        <div className="flex items-center gap-2">
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button variant="outline" size="sm">
-                <IconLayoutColumns />
-                <span className="hidden lg:inline">Columns</span>
-                <IconChevronDown />
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="end" className="w-56">
-              {table
-                .getAllColumns()
-                .filter((column) => typeof column.accessorFn !== "undefined" && column.getCanHide())
-                .map((column) => {
-                  return (
-                    <DropdownMenuCheckboxItem
-                      key={column.id}
-                      className="capitalize"
-                      checked={column.getIsVisible()}
-                      onCheckedChange={(value) => column.toggleVisibility(!!value)}
-                    >
-                      {column.id}
-                    </DropdownMenuCheckboxItem>
-                  );
-                })}
-            </DropdownMenuContent>
-          </DropdownMenu>
-          <Button asChild variant="outline" size="sm">
-            <Link to="/technologies/new">
-              <IconPlus />
-              <span className="hidden lg:inline">Add Section</span>
-            </Link>
-          </Button>
-        </div>
-      </div>
-      <TabsContent value="outline" className="relative flex flex-col gap-4 overflow-auto px-4 lg:px-6">
-        <div className="overflow-hidden rounded-lg border">
-          <DndContext
-            collisionDetection={closestCenter}
-            modifiers={[restrictToVerticalAxis]}
-            onDragEnd={handleDragEnd}
-            sensors={sensors}
-            id={sortableId}
-          >
-            <Table>
-              <TableHeader className="bg-muted sticky top-0 z-10">
-                {table.getHeaderGroups().map((headerGroup) => (
-                  <TableRow key={headerGroup.id}>
-                    {headerGroup.headers.map((header) => {
-                      return (
-                        <TableHead key={header.id} colSpan={header.colSpan}>
-                          {header.isPlaceholder ? null : (
-                            <div className={header.id === "select" ? "flex justify-center" : "flex align-middle"}>
-                              <div
-                                className={
-                                  header.column.getCanSort()
-                                    ? "cursor-pointer select-none flex self-center leading-[1.5rem]"
-                                    : " flex align-middle"
-                                }
-                                onClick={header.column.getToggleSortingHandler()}
-                                title={
-                                  header.column.getCanSort()
-                                    ? header.column.getNextSortingOrder() === "asc"
-                                      ? "Sort ascending"
-                                      : header.column.getNextSortingOrder() === "desc"
-                                        ? "Sort descending"
-                                        : "Clear sort"
-                                    : undefined
-                                }
-                              >
-                                {flexRender(header.column.columnDef.header, header.getContext())}
-                                <div>
-                                  {{
-                                    asc: <IconChevronUp />,
-                                    desc: <IconChevronDown />,
-                                  }[header.column.getIsSorted() as string] ?? null}
-                                </div>
-                              </div>
-                              {header.id === "actions" ? (
-                                <DropdownMenu>
-                                  <DropdownMenuTrigger asChild>
-                                    <Button
-                                      variant="ghost"
-                                      className="data-[state=open]:bg-muted text-muted-foreground flex size-8"
-                                      size="icon"
-                                    >
-                                      <IconDotsVertical />
-                                    </Button>
-                                  </DropdownMenuTrigger>
-                                  <DropdownMenuContent align="end" className="w-32">
-                                    <DropdownMenuItem
-                                      variant="destructive"
-                                      onClick={() => handleDeleteSelectedTechnologies()}
-                                    >
-                                      Delete selected
-                                    </DropdownMenuItem>
-                                  </DropdownMenuContent>
-                                </DropdownMenu>
-                              ) : null}
-                            </div>
-                          )}
-                        </TableHead>
-                      );
-                    })}
-                  </TableRow>
-                ))}
-              </TableHeader>
-              <TableBody className="**:data-[slot=table-cell]:first:w-8">
-                {isLoading ? (
-                  <TableRow>
-                    <TableCell colSpan={columns.length} className="h-24 text-center relative">
-                      <div className="flex justify-center items-center h-full w-full absolute top-0 left-0">
-                        <IconLoader className="animate-spin" />
-                      </div>
-                    </TableCell>
-                  </TableRow>
-                ) : (
-                  <>
-                    {table.getRowModel().rows?.length ? (
-                      <SortableContext items={dataIds} strategy={verticalListSortingStrategy}>
-                        {table.getRowModel().rows.map((row) => (
-                          <DraggableRow key={row.id} row={row} />
-                        ))}
-                      </SortableContext>
-                    ) : (
-                      <TableRow>
-                        <TableCell colSpan={columns.length} className="h-24 text-center">
-                          No results.
-                        </TableCell>
-                      </TableRow>
-                    )}
-                  </>
-                )}
-              </TableBody>
-            </Table>
-          </DndContext>
-        </div>
-        <div className="flex items-center justify-between px-4">
-          <div className="text-muted-foreground hidden flex-1 text-sm lg:flex">
-            {table.getFilteredSelectedRowModel().rows.length} of {table.getFilteredRowModel().rows.length} row(s)
-            selected.
-          </div>
-          <div className="flex w-full items-center gap-8 lg:w-fit">
-            <div className="hidden items-center gap-2 lg:flex">
-              <Label htmlFor="rows-per-page" className="text-sm font-medium">
-                Rows per page
-              </Label>
-              <Select
-                value={`${table.getState().pagination.pageSize}`}
-                onValueChange={(value) => {
-                  table.setPageSize(Number(value));
-                }}
-              >
-                <SelectTrigger size="sm" className="w-20" id="rows-per-page">
-                  <SelectValue placeholder={table.getState().pagination.pageSize} />
-                </SelectTrigger>
-                <SelectContent side="top">
-                  {[10, 20, 30, 40, 50].map((pageSize) => (
-                    <SelectItem key={pageSize} value={`${pageSize}`}>
-                      {pageSize}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            </div>
-            <div className="flex w-fit items-center justify-center text-sm font-medium">
-              Page {table.getState().pagination.pageIndex + 1} of {table.getPageCount()}
-            </div>
-            <div className="ml-auto flex items-center gap-2 lg:ml-0">
-              <Button
-                variant="outline"
-                className="hidden h-8 w-8 p-0 lg:flex"
-                onClick={() => table.setPageIndex(0)}
-                disabled={!table.getCanPreviousPage()}
-              >
-                <span className="sr-only">Go to first page</span>
-                <IconChevronsLeft />
-              </Button>
-              <Button
-                variant="outline"
-                className="size-8"
-                size="icon"
-                onClick={() => table.previousPage()}
-                disabled={!table.getCanPreviousPage()}
-              >
-                <span className="sr-only">Go to previous page</span>
-                <IconChevronLeft />
-              </Button>
-              <Button
-                variant="outline"
-                className="size-8"
-                size="icon"
-                onClick={() => table.nextPage()}
-                disabled={!table.getCanNextPage()}
-              >
-                <span className="sr-only">Go to next page</span>
-                <IconChevronRight />
-              </Button>
-              <Button
-                variant="outline"
-                className="hidden size-8 lg:flex"
-                size="icon"
-                onClick={() => table.setPageIndex(table.getPageCount() - 1)}
-                disabled={!table.getCanNextPage()}
-              >
-                <span className="sr-only">Go to last page</span>
-                <IconChevronsRight />
-              </Button>
-            </div>
-          </div>
-        </div>
-      </TabsContent>
-      <TabsContent value="past-performance" className="flex flex-col px-4 lg:px-6">
-        <div className="aspect-video w-full flex-1 rounded-lg border border-dashed"></div>
-      </TabsContent>
-      <TabsContent value="key-personnel" className="flex flex-col px-4 lg:px-6">
-        <div className="aspect-video w-full flex-1 rounded-lg border border-dashed"></div>
-      </TabsContent>
-      <TabsContent value="focus-documents" className="flex flex-col px-4 lg:px-6">
-        <div className="aspect-video w-full flex-1 rounded-lg border border-dashed"></div>
-      </TabsContent>
-    </Tabs>
-  );
-};
